requirements.txt
# Neural Pulse - Phase 1 Dependencies
# Runtime Verification Framework for SECA Attack Detection

# Core ML/DL Framework
# torch>=2.0.0
transformers>=4.45.0
accelerate>=0.24.0

# NLP Models
sentencepiece>=0.1.99
protobuf>=3.20.0

# LLM APIs (for SECA generation)
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# Signal Temporal Logic
rtamt>=0.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0

# Visualization (for Phase 2 analysis)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
scipy>=1.11.0
scikit-learn>=1.3.0

# JSON/YAML handling
pyyaml>=6.0.1

# Logging and Monitoring
tensorboard>=2.14.0

# Optional: For Nautilus K8s deployment
kubernetes>=28.0.0

# Development tools
jupyter>=1.0.0
ipywidgets>=8.1.0
black>=23.9.0
flake8>=6.1.0

# Note: For Llama-3-8B and DeBERTa models
# Requires HuggingFace authentication and model access
# Set HF_TOKEN environment variable or use `huggingface-cli login`
